% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cal_curve.R
\name{cal_curve}
\alias{cal_curve}
\title{Pulls out calibration samples and calibrates them}
\usage{
cal_curve(d, print.plot = TRUE, print.cal.stats = TRUE)
}
\description{
Takes a labeled `raw' data frame
(i.e., unprocessed but long-format and containing parsed labels for each well)
and returns  the same data frame, contiaining only samples (not standards),
plus columns for calibrated fluorescence (in units of concentration of fluorophore)
and associated error (plus p.value, etc).
\textbf{Note:} I probably want to re-form these things as a list
}

