% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_slopes.R
\name{calc_slopes}
\alias{calc_slopes}
\title{Calculate slopes & fit statistics}
\usage{
calc_slopes(dm, time_col = "Rtime", response_col = "fl",
  time_units = "hours", cols_to_ignore = c("std", "air.temp"))
}
\arguments{
\item{dm}{Data frame IN LONG FORMAT containing time & fluorescence data, as well as metadata (substrate, treatment, location, etc.)}

\item{time_col}{Column that contains the time, as POSIXct}

\item{response_col}{Column that contains the}

\item{time_units}{Units with which to express elapsed time (output)}

\item{cols_to_ignore}{Columns to ignore}
}
\description{
Calculates slopes, intercepts, standard errors & other fit statistics for each unique combination of other variables.
}

