% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_t_profile.R
\name{calc_t_profile}
\alias{calc_t_profile}
\title{Calculates t profile using patented Steen method}
\usage{
calc_t_profile(d, yvar = "fl", xvar = "elapsed", tvar = "temp")
}
\arguments{
\item{d}{Data frame containing columns with raw fluorescence, time or elapsed time, and temp}

\item{yvar}{Column name (as string) for response variable (fl)}

\item{xvar}{Column name (as string) for time variable (as POSIXct or numeric)}

\item{tvar}{Column name (as string) for temperature variable}
}
\description{
Calculates t_opt profile based on single-pot heating experiment
}
\details{
<<How this works>>
}

