% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calib_slope.R
\name{calib_slope}
\alias{calib_slope}
\title{Calibrated Slope}
\usage{
calib_slope(d, xvar = "conc.AMC.nM", yvar = "RFU")
}
\arguments{
\item{d}{- data frame}

\item{xvar}{- independent test variable}

\item{yvar}{- dependent test variable}
}
\value{
one row data frame containing the slope of the calibration curve
}
\description{
Calculate the slope of the calibration curve (in units of RFU per nM-AMC per L)
}

