% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Converts time into machine format}
\usage{
convert_time(d, date_col = NA, time_col = "time", format = "ymd_hm")
}
\arguments{
\item{date_col}{Column containing the date, as text <<or numeric?>>}

\item{time_col}{Column containing the time, as text <<or numeric?>>}

\item{format}{String indicating format of concatenated date and time}

\item{dm}{Main data frame}
}
\description{
Takes a data frame with time as strings. Adds a column called Rtime
}

