% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curve_plotr.R
\name{curve_plotr}
\alias{curve_plotr}
\title{Create plot of the Calibration Curve}
\usage{
curve_plotr(data, concentration.variable, fluorescence.variable, curvelabel,
  site.code)
}
\description{
Create plot of the Calibration Curve
}

