% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_plotr.R
\name{data_plotr}
\alias{data_plotr}
\title{Create faceted plot of data}
\usage{
data_plotr(data, datalabel = "Raw Data site", site.code, time.variable,
  fluorescence.variable, shape = "treatment", colour = "rep",
  fill = "rep")
}
\description{
Create faceted plot of data
}

