% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enza_plotr.R
\name{enza_plotr}
\alias{enza_plotr}
\title{Create plots of raw data, the calibration curve, and the calibrated activities}
\usage{
enza_plotr(plot.data = FALSE, plot.curve = FALSE, plot.v0 = FALSE, data,
  curve, activity, time.variable, fluorescence.variable, concentration.variable,
  site.code, datalabel = "Raw Data site_",
  curvelabel = "Calibration Curve For AMC, site ",
  v0label = "Calibrated v0, site ")
}
\arguments{
\item{plot.data}{}

\item{plot.curve}{}

\item{plot.v0}{}

\item{data}{}

\item{curve}{}

\item{activity}{}

\item{time.variable}{}

\item{fluorescence.variable}{}

\item{concentration.variable}{}

\item{site.code}{}

\item{datalabel}{}

\item{curvelabel}{}

\item{v0label}{}
}
\value{
created plots
}
\description{
Create plots of raw data, the calibration curve, and the calibrated activities
}

