% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enzalyze_reform.R
\name{enzalyze_reform}
\alias{enzalyze_reform}
\title{Melt and reshape}
\usage{
enzalyze_reform(d, id = c("rep", "treatment", "substrate"), var = "well",
  val = "fluorescence", the.date = NULL)
}
\arguments{
\item{d}{- input data frame (in long form)}

\item{id}{- vector of non-measured variables to be fed into the melt function}

\item{var}{- name of variable used to store measured variable names}

\item{val}{- variable used to store values}
}
\value{
returns the input data frame with two additional columns; one for the date-time object, and one for the calculated elapsed time (as a numeric vector)
}
\description{
Melt the data frame, calculate, and insert elapsed times into a long-form data frame
}

