% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm_stats.R
\name{lm_stats}
\alias{lm_stats}
\title{Returns statistics for linear regressions (e.g. slope, intercept, error, etc)}
\usage{
lm_stats(d, xvar, yvar)
}
\arguments{
\item{d}{Data frame containing at least two numeric columns}

\item{xvar}{Numeric column to be used as the independant variable for the regression}

\item{yvar}{Numeric column to be used as the dependant variable for the regression}
}
\value{
Returns a one-row data frame containing (at present): slope, intercept, slope standard error, intercept standard error, p value, r-squared, and number of points.
}
\description{
Returns statistics for linear regressions (e.g. slope, intercept, error, etc)
}
\details{
Under the assumption that there is an existing correlation between our scalar, dependent variable and our independent variable, lm_stat models the relationship by fitting a linear regression to a scatterplot of the data
}

