% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_names.R
\name{parse_names}
\alias{parse_names}
\title{Parses names of samples (or stds)}
\usage{
parse_names(d, name.col = "sample.name", separator = ";")
}
\description{
Parses names of samples
}
\details{
Explain rules here (; separation, 2 elements)
}

