% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_numeric.R
\name{parse_numeric}
\alias{parse_numeric}
\title{Parse numeric values and assign units}
\usage{
parse_numeric(x)
}
\arguments{
\item{x}{A character *vector* of numeric values, with units.}
}
\description{
Parse numeric values and assign units
}
\details{
The units must remain the same along the vector
}

