% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_biotek.R
\name{read_biotek}
\alias{read_biotek}
\title{Reads .csv output from Sean's Biotek
Note that I don't really know what the output from that instrument really looks like
For now I'll use a csv file with headers as well addresses (A1, A2, etc)}
\usage{
read_biotek(fn = NA, temp.var.name = "T..365.450",
  time.var.name = "Time..days.", temp.units = "oC", time.units = "days")
}
\description{
Reads .csv output from Sean's Biotek
Note that I don't really know what the output from that instrument really looks like
For now I'll use a csv file with headers as well addresses (A1, A2, etc)
}

