% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_long.R
\name{read_long}
\alias{read_long}
\title{Read raw data set as a data frame}
\usage{
read_long(x)
}
\arguments{
\item{x}{- raw dataset in textfile form}
}
\value{
data frame
}
\description{
Take in a raw data set, read as a data frame
}

