% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_plate_setup.R
\name{read_plate_setup}
\alias{read_plate_setup}
\title{Reads in plate setup worksheets}
\usage{
read_plate_setup(d)
}
\description{
Parses column names for tabular fluorescence data (where rows indicate separate sampling times)
}

