% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_NLS.R
\name{safe_NLS}
\alias{safe_NLS}
\title{Calculates exponential models safely}
\usage{
safe_NLS(df, xvar = "time", yvar = "relative.ion.count")
}
\arguments{
\item{df}{A data frame}

\item{xvar}{The x variable for the exponential model}

\item{yvar}{The y variable for the exponential model}
}
\description{
Calculates exponential models safely
}
\details{
**Works only if x variable is named time and y variable is named relative.ion.count**. THis is because of the way generate_exp_guess works, must fix.
}
