% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_NLS_pred.R
\name{safe_NLS_pred}
\alias{safe_NLS_pred}
\title{Generates nls predictions}
\usage{
safe_NLS_pred(mod, domain)
}
\arguments{
\item{mod}{is a nls model (or any object for which there is a predict method)}

\item{domain}{of the data, as a two-element vector. Will project these onto a 100-element grid}
}
\description{
Generates nls predictions
}
