% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_name_parser.R
\name{sample_name_parser}
\alias{sample_name_parser}
\title{Parses sample names. Right now takes a data frame of}
\usage{
sample_name_parser(d, sample_name_col = "sample.name",
  col_names = c("std.sample", "conc", "fluorophore", "buffer.sample"))
}
\arguments{
\item{names_vec}{A vector of names}
}
\description{
Parses sample names. Right now takes a data frame of
}

