% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_biotek.R
\name{tidy_biotek}
\alias{tidy_biotek}
\title{Puts biotek files into tidy form}
\usage{
tidy_biotek(x, n, d, temp.col = "temp", time.col = "Time..days.")
}
\arguments{
\item{x}{- file name of data set (.csv)}

\item{n}{- number of rows to read as data frame from text file}

\item{d}{- horizontal data frame}
}
\value{
***put something here***
}
\description{
Formats raw biotek files into long form with standardized column headings
}

