% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uncalib_slope.R
\name{uncalib_slope}
\alias{uncalib_slope}
\title{Uncalibrated Slope}
\usage{
uncalib_slope(d, id.var = c("rep", "treatment", "substrate"),
  xvar = "elapsed", yvar = "RFU")
}
\arguments{
\item{d}{- data frame to be fed to plyr and lm_stats functions}

\item{id.var}{- variables to split input data frame by before applying lm_stats for the slope}

\item{xvar}{- independent variable for the linear model}

\item{yvar}{- dependent variable for the linear model}
}
\value{
data frame containing the linear regression statistics for each combination of id-variables
}
\description{
Calculate the linear regression statistics, from the uncalibrated data, for each unique
combination of id-variables, including the uncalibrated slopes (in units of RFU/hr)
}

