% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Activity of enzymes}
\usage{
activity(uncal, cal, .id = c("rep", "treatment", "substrate"),
  .var = "well", .val = "fluorescence", .the.date = NULL,
  .id.var = c("rep", "treatment", "substrate"), .xvar = "elapsed",
  .yvar = "RFU", d, xvar = "conc.AMC.nM")
}
\arguments{
\item{uncal}{Raw dataset, of uncalibrated activity, in textfile form}

\item{cal}{Raw dataset containing the calibration curve}

\item{.id}{Vector of non-measured variables to melt the data frame by}

\item{.var}{Name of variable used to store measured variable names}

\item{.val}{Variable used to store values}

\item{.the.date}{The date the data was gathered (default will print the current date)}

\item{.id.var}{Vector of variables by which to split input data frame by, before applying lm_stats to calculate slope values}

\item{.xvar}{Independent variable for the uncalibrated linear model}

\item{.yvar}{Dependent variable for the linear model (same for uncalibrated & calibrated models)}

\item{xvar}{Independent variable for the calibration curve linear model}
}
\value{
One-row data frame containing the activity of enzymes
}
\description{
Calculates enzymatic activity from a raw, uncalibrated dataset and a calibration curve
}

