\name{enzalyze-package}
\alias{enzalyze-package}
\alias{enzalyze}
\docType{package}
\title{
Measure enzyme activities
~~ enzalyze ~~
}
\description{

~~ Within enzalyze are all the tools to calculate enzyme activities.  Beginning with melting data files, of various formats, into long form and reading them as a data frame.  Then by taking that data frame, extracting the statistics from a linear regression, and utilizing that data along with data from a calibration curve (also within the package) we can calculate true enzyme activity ~~

}
\details{
\tabular{ll}{
Package: \tab enzalyze\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-03\cr
License: \tab What license is it under?\cr
}
~~ DELETE AFTER YOU ADD OTHER FUNCTION DESCRIPTIONS - An overview of how to use the package, including the most important functions ~~
~~ lm_stats : create a linear model between a designated dependent and independent variable, from a long-form data frame, and return the regression statistics (i.e slope, intercept, error, etc) ~~
}
\author{
Asst. Professor Andrew Steen
(with help from Josh Gurka)

Maintainer: Josh Gurka <jgurka@vols.utk.edu>
~~ Andrew Steen & Josh Gurka ~~
}
\references{
~~ Advanced R, Hadley Wickham ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
