% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enzalyze_reform.R
\name{enzalyze_reform}
\alias{enzalyze_reform}
\title{Name substrates & calculate elapsed time}
\usage{
enzalyze_reform(d, .labels = c("Arg-AMC", "Gly-AMC", "Leu-AMC", "Pyr-AMC",
  "GlyGlyArg-AMC"), the.date = NULL)
}
\arguments{
\item{d}{- input data frame (in long form)}

\item{.labels}{- character string replacement for numeric labels of substrates}

\item{the.date}{- date.  Default results in the system date being called and used}
}
\value{
returns the input data frame with two additional columns; one for the date-time object, and one for the calculated elapsed time (as a numeric vector)
}
\description{
Relabel the substrates in the data frame, then calculate and insert elapsed times into a new column of the data frame
}

