% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uncalib_slope.R
\name{uncalib_slope}
\alias{uncalib_slope}
\title{Uncalibrated Slope}
\usage{
uncalib_slope(d, id.var = c("rep", "treatment", "substrate"),
  time.var = "elapsed", fluorescence = "RFU")
}
\arguments{
\item{d}{- data frame to be fed to plyr and lm_stats functions}

\item{id.var}{- variables to split input data frame by before applying the `lm_stats` function}

\item{time.var}{- independent variable for the linear model, representing the time label for the model (i.e "minutes", "seconds", "elapsed")}

\item{fluorescence}{- dependent variable for the linear model, representing the units of fluorescence from data collection}
}
\value{
data frame containing the linear regression statistics for each combination of id-variables
}
\description{
Calculate the linear regression statistics, from the uncalibrated data, for each unique
combination of id-variables, including the uncalibrated slopes (in units of RFU/hr)
}

